+++ specs/009-unsigned-ledger-builder/tasks.md
@@ -29,10 +29,10 @@ No separate foundational implementation commit. The acyclic builder boundary is 
 
 **Independent Test**: focused Shelley transaction unit specs plus `./gate.sh`.
 
-- [ ] T010 [US1] Add/adjust RED tests in `lib/unit/test/unit/Cardano/Wallet/Shelley/TransactionSpec.hs` and/or `lib/unit/test/unit/Cardano/Wallet/Shelley/TransactionLedgerSpec.hs` so the desired ledger-cert `mkUnsignedTx` wrapper path is required and the pre-change code fails to compile or fails the test.
-- [ ] T011 [US1] Implement the acyclic shared unsigned ledger-builder boundary in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction/Build.hs` or a new sibling module, updating `Transaction.Ledger` to reexport/delegate without importing cycles.
-- [ ] T012 [US1] Migrate `mkUnsignedTransaction` and `mkUnsignedTx` in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs` off `Cardano.createTransactionBody` / `Cardano.TxBodyContent`, preserving existing `Cardano.TxBody` return interfaces where still required.
-- [ ] T013 [US2] Switch the remaining certificate call sites in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs` to ledger-native helpers, delete `lib/wallet/src/Cardano/Wallet/Transaction/Delegation.hs` and `lib/wallet/src/Cardano/Wallet/Transaction/Voting.hs`, and prune both exposed modules from `lib/wallet/cardano-wallet.cabal`.
+- [X] T010 [US1] Add/adjust RED tests in `lib/unit/test/unit/Cardano/Wallet/Shelley/TransactionSpec.hs` and/or `lib/unit/test/unit/Cardano/Wallet/Shelley/TransactionLedgerSpec.hs` so the desired ledger-cert `mkUnsignedTx` wrapper path is required and the pre-change code fails to compile or fails the test.
+- [X] T011 [US1] Implement the acyclic shared unsigned ledger-builder boundary in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction/Build.hs` or a new sibling module, updating `Transaction.Ledger` to reexport/delegate without importing cycles.
+- [X] T012 [US1] Migrate `mkUnsignedTransaction` and `mkUnsignedTx` in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs` off `Cardano.createTransactionBody` / `Cardano.TxBodyContent`, preserving existing `Cardano.TxBody` return interfaces where still required.
+- [X] T013 [US2] Switch the remaining certificate call sites in `lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs` to ledger-native helpers, delete `lib/wallet/src/Cardano/Wallet/Transaction/Delegation.hs` and `lib/wallet/src/Cardano/Wallet/Transaction/Voting.hs`, and prune both exposed modules from `lib/wallet/cardano-wallet.cabal`.
 
 T010-T013 fold into one commit with subject `feat(5285): migrate unsigned Shelley tx builder` and trailer `Tasks: T010, T011, T012, T013`.
